+++ src/doc/changelog.md
@@ -6,6 +6,7 @@
 - Add scale line feature
 - Add swipe feature
 - Add heatmap layer
+- Add heatmap KML layer
 
 NOTE: This WireCloud widget is based on Wirecloud/ol3-map-widget created by
 CoNWeT Lab., Universidad Politecnica de Madrid and Future Internet Consulting